+++ src/tests.rs
@@ -71,7 +71,7 @@ fn check_struct_sizes() {
     assert_eq!(size_of::<Scope>(), 24);
     assert_eq!(
         size_of::<FnPtr>(),
-        32 - if cfg!(feature = "no_function") {
+        48 - if cfg!(feature = "no_function") {
             WORD_SIZE
         } else {
             0